feat: creating role crud

+++ BackEnd/MyMechanic/src/main/java/com/mechtech/MyMechanic/web/dto/serviceorder/ServiceOrderEmployeeResponseDto.java
@@ -12,5 +12,5 @@ import java.math.BigDecimal;
 public class ServiceOrderEmployeeResponseDto {
     private Long employeeId;
     private String employeeName;
-    private int commissionPercentage;
+    private BigDecimal commissionPercentage;
 }
